laatste wijzigingen omtrent auth

+++ src/main/java/com/iprwc/jpgshop/controller/ProductController.java
@@ -1,7 +1,12 @@
 package com.iprwc.jpgshop.controller;
 
+import com.iprwc.jpgshop.config.JwtFilter;
+import com.iprwc.jpgshop.config.JwtToken;
 import com.iprwc.jpgshop.entity.Product;
+import com.iprwc.jpgshop.entity.User;
 import com.iprwc.jpgshop.service.ProductService;
+import com.iprwc.jpgshop.service.UserService;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -11,15 +16,31 @@ import java.util.List;
 @RequestMapping("/api/products")
 public class ProductController {
 
+    private final JwtFilter jwtFilter;
+    private final UserService userService;
+    private final JwtToken jwtToken;
     ProductService productService;
 
-    public ProductController(ProductService productService) {
+    public ProductController(JwtFilter jwtFilter, UserService userService, JwtToken jwtToken, ProductService productService) {
+        this.jwtFilter = jwtFilter;
+        this.userService = userService;
+        this.jwtToken = jwtToken;
         this.productService = productService;
     }
 
     @PostMapping("/createProduct")
-    public void createProduct(@RequestBody Product product) {
-        productService.createProduct(product);
+    public void createProduct(HttpServletRequest request, @RequestBody Product product) {
+        String token = jwtFilter.parseJwt(request);
+        System.out.println(token);
+        if (token != null && !token.isBlank()) {
+            User user = userService.findUserByToken(token);
+            System.out.println(user);
+            if (user.getEmail().equals("test@test.com")) {
+                productService.createProduct(product);
+                return;
+            }
+        }
+        throw new RuntimeException("Error: could not create product.");
     }
 
     @GetMapping("/getProducts")
@@ -29,12 +50,28 @@ public class ProductController {
     }
 
     @PutMapping("/updateProduct/{productId}")
-    public void updateProduct(@PathVariable Long productId, @RequestBody Product product) {
-        productService.updateProduct(productId, product);
+    public void updateProduct(HttpServletRequest request, @PathVariable Long productId, @RequestBody Product product) {
+        String token = jwtFilter.parseJwt(request);
+        if (token != null && !token.isBlank()) {
+            User user = userService.findUserByToken(token);
+            if (user.getEmail().equals("test@test.com")) {
+                productService.updateProduct(productId, product);
+                return;
+            }
+        }
+        throw new RuntimeException("Error: could not update product.");
     }
 
     @DeleteMapping("/deleteProduct/{productId}")
-    public void deleteProduct(@PathVariable Long productId) {
-        productService.deleteProductById(productId);
+    public void deleteProduct(HttpServletRequest request, @PathVariable Long productId) {
+        String token = jwtFilter.parseJwt(request);
+        if (token != null && !token.isBlank()) {
+            User user = userService.findUserByToken(token);
+            if (user.getEmail().equals("test@test.com")) {
+                productService.deleteProductById(productId);
+                return;
+            }
+        }
+        throw new RuntimeException("Error: could not delete product.");
     }
 }
